Fix regression: Add <input> back to <form>...
My tidying up apparently led to a regression that removed the injected
<input> from <form> tags.

Doh! Evidently this module needs many more tests. :) Ironically it was
my HTML test program that detected this... Perhaps I can work on adding
additional tests in the future. Sorry for the confusion and trouble! I
can't believe I missed this. I'm actually sure that I caught this
earlier, but it must have gotten lost in rebasing or something... Who
knows.